pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "prompt>=0.4.1",
    "PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

# ИСПРАВЛЯЕМ ЭТУ СТРОКУ:
[project.scripts]
gendiff = "gendiff.__main__:main"

[tool.ruff]
line-length = 80

[lint.per-file-ignores]
  # init modules can contain the local imports, logic, unused imports
"__init__.py" = ["F401"]

[lint]
preview = true
select = ["E", "F", "I", "C90"]